backend/requirements.txt
Django==3.2.6
psycopg2-binary==2.9.1
django-cors-headers==3.10.1
djangorestframework
# django-geofield==0.0.2